fix : prefilled data on edit click

+++ Ticket Web application/angular/src/app/tickets/tickets.component.ts
@@ -26,7 +26,7 @@ export class TicketsComponent implements OnInit {
   dataSource = new MatTableDataSource<Ticket>();
   public ticket: Array<Object>;
 
-  currentUser = JSON.parse(sessionStorage.getItem('auth-user')!)
+  currentUser = JSON.parse(sessionStorage.getItem('auth-user')!);
 
   constructor(
     private _liveAnnouncer: LiveAnnouncer,
@@ -35,12 +35,12 @@ export class TicketsComponent implements OnInit {
   ) {
     this.fetchTicket();
   }
-  
+
   handleEdit = (title: String, desc: String, id: Number) => {
     this.dialog.open(UpdateTicketComponent, {
       data: {
-        title_title: title,
-        title_desc: desc,
+        ticket_title: title,
+        ticket_desc: desc,
         editId: id,
       },
     });
@@ -53,18 +53,19 @@ export class TicketsComponent implements OnInit {
       },
     });
   }
-  
+
   @ViewChild(MatPaginator) paginator: MatPaginator;
   @ViewChild(MatSort) sort: MatSort;
-  
+
   ngOnInit(): void {
     this.fetchTicket();
   }
 
   handleDelete = (id: Number) => {
-    this.ticketService.deleteTicketService(id).subscribe((data) => {
-      this.fetchTicket();
-    });
+    confirm('Are you sure?') &&
+      this.ticketService.deleteTicketService(id).subscribe((data) => {
+        this.fetchTicket();
+      });
   };
 
   fetchTicket = () => {
